--- .github/PULL_REQUEST_TEMPLATE.md
@@ -1,7 +1,0 @@
-Because
-
-- (write the reason why we need to consider this PR in a list)
-
-This commit
-
-- (write the summary of all commits in this PR in a list)
