--- .github/SECURITY.md
@@ -1,3 +1,0 @@
-If you discover a security issue in this repository, please contact us through security@instill.tech.
-
-Thanks for helping make the codebase safe for everyone.
