+++ .github/labeler.yml
@@ -1,0 +1,2 @@
+instill core:
+  - "**"
