--- .github/workflows/add-issue-to-prj.yml
@@ -1,14 +1,0 @@
-name: Add Issue to Project
-
-on:
-  issues:
-    types:
-      - opened
-
-jobs:
-  track_issue:
-    uses: instill-ai/meta/.github/workflows/add-issue-to-prj.yml@main
-    with:
-      project_number: 12 # Instill Base project
-    secrets:
-      botGitHubToken: ${{ secrets.botGitHubToken }}
